Update LUT count comment in ALSRU
The number of LUTs necessary to compute the zero flag is relatively easy
to predict, as the number of inputs a LUT tree can handle doesn't depend
on the shape of the tree — an additional 4LUT always lets you handle 3
more inputs.

I haven't actually tested that Yosys+ABC achieves this, but since
prjunnamed does, I'd be very surprised if ABC doesn't.